# openapi.yaml
openapi: 3.0.3
info:
  title: API Tablero Cooptech
  version: 1.0.0
  description: API REST del Tablero de Mando Cooptech. Single-tenant, uso interno.
    Autenticación delegada al servicio de identidad central (Bearer JWT).
servers:
- url: /api/v1
security:
- bearerAuth: []
...
paths:
...
  /tareas:
    get:
      tags:
      - tareas
      summary: Listar tareas
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tarea'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - tareas
      summary: Crear tarea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarea'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /tareas/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - tareas
      summary: Obtener por id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarea'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - tareas
      summary: Actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarea'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Validation'
    delete:
      tags:
      - tareas
      summary: Eliminar
      responses:
        '204':
          description: Eliminado
        '404':
          $ref: '#/components/responses/NotFound'
...
  /leads:
    get:
      tags:
      - leads
      summary: Listar leads
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - leads
      summary: Crear lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadInput'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '422':
          $ref: '#/components/responses/Validation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /leads/{id}:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - leads
      summary: Obtener por id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - leads
      summary: Actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadInput'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lead'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Validation'
    delete:
      tags:
      - leads
      summary: Eliminar
      responses:
        '204':
          description: Eliminado
        '404':
          $ref: '#/components/responses/NotFound'
...
  /leads/{id}/actividades:
    parameters:
    - $ref: '#/components/parameters/Id'
    get:
      tags:
      - leads
      summary: Actividades del lead
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrmActividad'
    post:
      tags:
      - leads
      summary: Registrar actividad (visita/videollamada/evento)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrmActividadInput'
      responses:
        '201':
          description: Creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrmActividad'
  /leads/{id}/ganar:
    parameters:
    - $ref: '#/components/parameters/Id'
    post:
      tags:
      - leads
      summary: Marcar ganado y crear el proyecto (hand-off CRM -> Kanban)
      responses:
        '200':
          description: Lead ganado y proyecto creado
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead:
                    $ref: '#/components/schemas/Lead'
                  proyecto:
                    $ref: '#/components/schemas/Proyecto'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: *id002
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 50
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Validation:
      description: Datos inválidos
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: *id004
            message: *id004
    Pagination:
      type: object
      properties:
        page: *id002
        pageSize: *id002
        total: *id002
...
    Proyecto:
      type: object
      properties:
        id: *id002
        nombre: *id004
        enfoque: &id014
          type: string
          enum: &id023
          - Organización
          - Electrónica
          - Desarrollo web
          - Comercialización
          - Operación
          nullable: true
        estado: &id015
          type: string
          enum:
          - activo
          - pausado
          - cerrado
        leadId: &id016
          type: integer
          nullable: true
...
    Tarea:
      type: object
      properties:
        id: *id002
        proyectoId: &id017
          type: integer
          nullable: true
        titulo: *id004
        descripcion: &id018
          type: string
          nullable: true
        kanbanCol: &id019
          type: string
          enum:
          - backlog
          - todo
          - doing
          - done
        prioridad: &id020
          type: string
          enum:
          - baja
          - media
          - alta
          - urgente
        orden: *id002
        responsablesIds: &id021
          type: array
          items: *id002
        tagsIds: &id022
          type: array
          items: *id002
    TareaInput:
      type: object
      properties:
        proyectoId: *id017
        titulo: *id004
        descripcion: *id018
        kanbanCol: *id019
        prioridad: *id020
        orden: *id002
        responsablesIds: *id021
        tagsIds: *id022
...
    Lead:
      type: object
      properties:
        id: *id002
        organizacion: *id004
        contactoNombre: &id035
          type: string
          nullable: true
        telefono: &id036
          type: string
          nullable: true
        email: &id037
          type: string
          nullable: true
        ciudad: &id038
          type: string
          nullable: true
        fechaPrimerContacto: &id039
          type: string
          format: date
          nullable: true
        ownerId: &id040
          type: integer
          nullable: true
        etapa: &id041
          type: string
          enum:
          - contacto
          - visita_agendada
          - visita_realizada
          - propuesta
          - negociacion
          - trial
          - ganado
          - perdido
        valorEstimadoUsd: &id042
          type: number
          nullable: true
        montoFacturadoUsd: &id043
          type: number
          nullable: true
        productos: &id044
          type: array
          items: *id004
        presupuestoEstado: &id034
          type: object
          nullable: true
        presupuestoAguaEstado: *id034
        coopcloudEstado: *id034
        coopcloudCostoMensual: &id045
          type: number
          nullable: true
        notas: &id046
          type: string
          nullable: true
        fuente: &id047
          type: string
          nullable: true
    LeadInput:
      type: object
      properties:
        organizacion: *id004
        contactoNombre: *id035
        telefono: *id036
        email: *id037
        ciudad: *id038
        fechaPrimerContacto: *id039
        ownerId: *id040
        etapa: *id041
        valorEstimadoUsd: *id042
        montoFacturadoUsd: *id043
        productos: *id044
        presupuestoEstado: *id034
        presupuestoAguaEstado: *id034
        coopcloudEstado: *id034
        coopcloudCostoMensual: *id045
        notas: *id046
        fuente: *id047
...
    CrmActividad:
      type: object
      properties:
        id: *id002
        leadId:
          type: integer
          nullable: true
        colaboradorId:
          type: integer
          nullable: true
        tipo:
          type: string
          enum:
          - visita
          - videollamada
          - evento
        fecha: *id003
        notas:
          type: string
          nullable: true
    CrmActividadInput:
      type: object
      properties:
        colaboradorId:
          type: integer
          nullable: true
        tipo:
          type: string
          enum:
          - visita
          - videollamada
          - evento
        fecha: *id003
        notas:
          type: string
          nullable: true
      required:
      - tipo
      - fecha
...